Move download all content to total attention
